chore: type safe api.
