+++ util/movement-aptos/movement-aptos-core/src/movement_aptos/runtime.rs
@@ -1,3 +1,11 @@
+pub mod delegated;
+pub mod native;
+pub mod tokio_test;
+
+pub use delegated::*;
+pub use native::*;
+pub use tokio_test::*;
+
 use std::fmt::Debug;
 
 /// Errors thrown when attempting to use a runtime.
@@ -7,14 +15,6 @@ pub enum RuntimeError {
 	Internal(#[source] Box<dyn std::error::Error + Send + Sync>),
 	#[error("requested Movement Aptos Runtime is unavailable: {0}")]
 	Unavailable(#[source] Box<dyn std::error::Error + Send + Sync>),
-}
-
-/// Returns true if the current runtime is multithreaded.
-fn is_multithreaded_runtime() -> bool {
-	std::panic::catch_unwind(|| {
-		tokio::task::block_in_place(|| {});
-	})
-	.is_ok()
 }
 
 /// Trait for a runtime that can be used to run [MovementAptos].
@@ -28,63 +28,15 @@ pub trait Runtime: Sized + Clone + Debug + Send + Sync + 'static {
 	fn create_global_rayon_pool() -> bool;
 }
 
-/// Tokio test runtime.
-#[derive(Debug, Clone)]
-pub struct TokioTest;
-
-impl Runtime for TokioTest {
-	/// Try to create a new runtime.
-	fn try_new() -> Result<Self, RuntimeError> {
-		if !is_multithreaded_runtime() {
-			return Err(RuntimeError::Unavailable(Box::new(std::io::Error::new(
-				std::io::ErrorKind::Other,
-				"Tokio test runtime is not multithreaded use #[tokio::test(flavor = \"multi_thread\")] instead",
-			))));
-		}
-
-		Ok(Self)
-	}
-
-	/// Whether to create a global rayon pool.
-	fn create_global_rayon_pool() -> bool {
-		false
-	}
+/// Returns true if the current runtime is multithreaded.
+fn is_multithreaded_runtime() -> bool {
+	std::panic::catch_unwind(|| {
+		tokio::task::block_in_place(|| {});
+	})
+	.is_ok()
 }
 
-/// Native runtime refers to a runtime where the global rayon pool is created within the runner.
-#[derive(Debug, Clone)]
-pub struct Native;
-
-impl Runtime for Native {
-	/// Try to create a new runtime.
-	///
-	/// There are no restrictions on the surrounding environment here.
-	fn try_new() -> Result<Self, RuntimeError> {
-		Ok(Self)
-	}
-
-	/// Whether to create a global rayon pool.
-	fn create_global_rayon_pool() -> bool {
-		true
-	}
-}
-
-/// Delegated runtime refers to a runtime where the global rayon pool is created outside of the runner.
-///
-/// This is useful when we may be calling from other tasks, i.e., not at the main thread.
-#[derive(Debug, Clone)]
-pub struct Delegated;
-
-impl Runtime for Delegated {
-	/// Try to create a new runtime.
-	///
-	/// There are no restrictions on the surrounding environment here.
-	fn try_new() -> Result<Self, RuntimeError> {
-		Ok(Self)
-	}
-
-	/// Whether to create a global rayon pool.
-	fn create_global_rayon_pool() -> bool {
-		false
-	}
+/// Returns true if the current runtime is a tokio runtime.
+fn is_in_tokio_runtime() -> bool {
+	tokio::runtime::Handle::try_current().is_ok()
 }
